# Speedup Travis CI jobs