--- a/integration-tests/src/test/java/org/apache/druid/tests/hadoop/ITHadoopIndexTest.java
+++ b/integration-tests/src/test/java/org/apache/druid/tests/hadoop/ITHadoopIndexTest.java
@@ -25,14 +25,14 @@ import org.apache.druid.java.util.common.logger.Logger;
 import org.apache.druid.testing.IntegrationTestingConfig;
 import org.apache.druid.testing.guice.DruidTestModuleFactory;
 import org.apache.druid.testing.utils.RetryUtil;
+import org.apache.druid.tests.TestNGGroup;
 import org.apache.druid.tests.indexer.AbstractIndexerTest;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Guice;
 import org.testng.annotations.Test;
 
-import java.util.concurrent.Callable;
-
+@Test(groups = TestNGGroup.HADOOP_INDEX)
 @Guice(moduleFactory = DruidTestModuleFactory.class)
 public class ITHadoopIndexTest extends AbstractIndexerTest
 {
@@ -77,15 +77,7 @@ public class ITHadoopIndexTest extends AbstractIndexerTest
       LOG.info("TaskID for loading index task %s", taskID);
       indexer.waitUntilTaskCompletes(taskID, 10000, 120);
       RetryUtil.retryUntil(
-          new Callable<Boolean>()
-          {
-            @Override
-            public Boolean call()
-            {
-              return coordinator.areSegmentsLoaded(BATCH_DATASOURCE);
-
-            }
-          },
+          () -> coordinator.areSegmentsLoaded(BATCH_DATASOURCE),
           true,
           20000,
           10,