--- a/integration-tests/src/test/java/org/apache/druid/tests/indexer/ITAppenderatorDriverRealtimeIndexTaskTest.java
+++ b/integration-tests/src/test/java/org/apache/druid/tests/indexer/ITAppenderatorDriverRealtimeIndexTaskTest.java
@@ -28,6 +28,7 @@ import org.apache.druid.java.util.common.logger.Logger;
 import org.apache.druid.testing.clients.EventReceiverFirehoseTestClient;
 import org.apache.druid.testing.guice.DruidTestModuleFactory;
 import org.apache.druid.testing.utils.ServerDiscoveryUtil;
+import org.apache.druid.tests.TestNGGroup;
 import org.joda.time.DateTime;
 import org.testng.annotations.Guice;
 import org.testng.annotations.Test;
@@ -43,6 +44,7 @@ import java.util.Map;
 /**
  * See {@link AbstractITRealtimeIndexTaskTest} for test details.
  */
+@Test(groups = TestNGGroup.REALTIME_INDEX)
 @Guice(moduleFactory = DruidTestModuleFactory.class)
 public class ITAppenderatorDriverRealtimeIndexTaskTest extends AbstractITRealtimeIndexTaskTest
 {
@@ -57,11 +59,10 @@ public class ITAppenderatorDriverRealtimeIndexTaskTest extends AbstractITRealtim
   }
 
   @Override
-  public void postEvents() throws Exception
+  void postEvents() throws Exception
   {
     final ServerDiscoverySelector eventReceiverSelector = factory.createSelector(EVENT_RECEIVER_SERVICE_NAME);
     eventReceiverSelector.start();
-    BufferedReader reader = null;
     InputStreamReader isr;
     try {
       isr = new InputStreamReader(
@@ -72,8 +73,7 @@ public class ITAppenderatorDriverRealtimeIndexTaskTest extends AbstractITRealtim
     catch (Exception e) {
       throw new RuntimeException(e);
     }
-    try {
-      reader = new BufferedReader(isr);
+    try (BufferedReader reader = new BufferedReader(isr)) {
       ServerDiscoveryUtil.waitUntilInstanceReady(eventReceiverSelector, "Event Receiver");
       // Use the host from the config file and the port announced in zookeeper
       String host = config.getMiddleManagerHost() + ":" + eventReceiverSelector.pick().getPort();
@@ -102,7 +102,7 @@ public class ITAppenderatorDriverRealtimeIndexTaskTest extends AbstractITRealtim
         }
         String event = StringUtils.replace(line, TIME_PLACEHOLDER, EVENT_FMT.print(dt));
         LOG.info("sending event: [%s]\n", event);
-        Collection<Map<String, Object>> events = new ArrayList<Map<String, Object>>();
+        Collection<Map<String, Object>> events = new ArrayList<>();
         events.add(this.jsonMapper.readValue(event, JacksonUtils.TYPE_REFERENCE_MAP_STRING_OBJECT));
         int eventsPosted = client.postEvents(events, this.jsonMapper, MediaType.APPLICATION_JSON);
         if (eventsPosted != events.size()) {
@@ -112,7 +112,7 @@ public class ITAppenderatorDriverRealtimeIndexTaskTest extends AbstractITRealtim
         try {
           Thread.sleep(DELAY_BETWEEN_EVENTS_SECS * 1000);
         }
-        catch (InterruptedException ex) {
+        catch (InterruptedException ignored) {
           /* nothing */
         }
         dtLast = dt; // latest timestamp
@@ -124,7 +124,6 @@ public class ITAppenderatorDriverRealtimeIndexTaskTest extends AbstractITRealtim
       throw new RuntimeException(e);
     }
     finally {
-      reader.close();
       eventReceiverSelector.stop();
     }
   }