@himanshug 2019-08-21
Thanks, this is awesome . Travis build now runs faster than teamcity and other checks :)